+++ common-data-fetchers/src/main/java/io/strimzi/admin/common/data/fetchers/AdminClientWrapper.java
@@ -2,8 +2,8 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.admin.kafka.admin;
 
+package io.strimzi.admin.common.data.fetchers;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
