+++ common-data-fetchers/src/main/java/io/strimzi/admin/common/data/fetchers/model/Types.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.admin.kafka.admin.model;
+package io.strimzi.admin.common.data.fetchers.model;
 
 import java.util.List;
