+++ kafka-admin/src/main/java/io/strimzi/admin/kafka/admin/handlers/TopicCreateHandler.java
@@ -4,20 +4,15 @@
  */
 package io.strimzi.admin.kafka.admin.handlers;
 
-import io.strimzi.admin.kafka.admin.AdminClientWrapper;
-import io.strimzi.admin.kafka.admin.model.Types;
-import io.vertx.core.Promise;
+import io.strimzi.admin.common.data.fetchers.AdminClientWrapper;
+import io.strimzi.admin.common.data.fetchers.TopicOperations;
+import io.strimzi.admin.common.data.fetchers.model.Types;
 import io.vertx.core.Vertx;
-import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.handler.graphql.VertxDataFetcher;
-import io.vertx.kafka.admin.NewTopic;
-import org.apache.kafka.common.config.SaslConfigs;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -26,30 +21,9 @@ public class TopicCreateHandler extends CommonHandler {
 
     public static VertxDataFetcher createTopic(Map<String, Object> acConfig, Vertx vertx) {
         VertxDataFetcher<Types.Topic> dataFetcher = new VertxDataFetcher<>((environment, prom) -> {
-            RoutingContext rc = environment.getContext();
-            String token = rc.request().getHeader("Authorization");
-            if (token != null) {
-                if (token.startsWith("Bearer ")) {
-                    token = token.substring("Bearer ".length());
-                }
-                log.info("auth token is {}", token);
-                log.info(SaslConfigs.SASL_JAAS_CONFIG + "is org.apache.kafka.common.security.oauthbearer.OAuthBearerLoginModule required oauth.access.token=\"" + token + " \";");
-                acConfig.put(SaslConfigs.SASL_JAAS_CONFIG, "org.apache.kafka.common.security.oauthbearer.OAuthBearerLoginModule required oauth.access.token=\"" + token + "\";");
-            }
+            setOAuthToken(acConfig, environment);
+            AdminClientWrapper acw = createAdminClient(vertx, acConfig, prom);
 
-            AdminClientWrapper acw = new AdminClientWrapper(vertx, acConfig);
-            try {
-                acw.open();
-            } catch (Exception e) {
-                log.error(e);
-                if (acw != null) {
-                    acw.close();
-                }
-                prom.fail(e);
-                return;
-            }
-
-            NewTopic newKafkaTopic = new NewTopic();
             Types.NewTopic inputTopic = new Types.NewTopic();
 
             Map<String, Object> input = environment.getArgument("input");
@@ -68,45 +42,7 @@ public class TopicCreateHandler extends CommonHandler {
             inputTopic.setNumPartitions(Integer.parseInt(input.get("numPartitions").toString()));
             inputTopic.setReplicationFactor(Integer.parseInt(input.get("replicationFactor").toString()));
 
-            Map<String, String> config = new HashMap<>();
-            List<Types.NewTopicConfigEntry> configObject = inputTopic.getConfig();
-            configObject.forEach(item -> {
-                config.put(item.getKey(), item.getValue());
-            });
-
-            newKafkaTopic.setName(inputTopic.getName());
-            newKafkaTopic.setReplicationFactor(inputTopic.getReplicationFactor().shortValue());
-            newKafkaTopic.setNumPartitions(inputTopic.getNumPartitions());
-            if (config != null) {
-                newKafkaTopic.setConfig(config);
-            }
-
-            Promise createTopicPromise = Promise.promise();
-            acw.createTopic(Collections.singletonList(newKafkaTopic), res -> {
-                if (res.failed()) {
-                    log.error(res.cause());
-                    prom.fail(res.cause());
-                    acw.close();
-                } else {
-                    createTopicPromise.complete(res);
-                }
-            });
-
-            createTopicPromise.future().onComplete(ignore -> {
-                Types.Topic topic = new Types.Topic();
-                List<Types.ConfigEntry> newConf = new ArrayList<>();
-                inputTopic.getConfig().forEach(in -> {
-                    Types.ConfigEntry configEntry = new Types.ConfigEntry();
-                    configEntry.setKey(in.getKey());
-                    configEntry.setValue(in.getValue());
-                    newConf.add(configEntry);
-                });
-
-                topic.setConfig(newConf);
-                topic.setName(inputTopic.getName());
-                prom.complete(topic);
-                acw.close();
-            });
+            TopicOperations.createTopic(acw, prom, inputTopic);
         });
         return dataFetcher;
     }
