+++ kafka-admin/src/main/java/io/strimzi/admin/kafka/admin/handlers/TopicsDeleteHandler.java
@@ -4,11 +4,10 @@
  */
 package io.strimzi.admin.kafka.admin.handlers;
 
-import io.strimzi.admin.kafka.admin.AdminClientWrapper;
+import io.strimzi.admin.common.data.fetchers.AdminClientWrapper;
+import io.strimzi.admin.common.data.fetchers.TopicOperations;
 import io.vertx.core.Vertx;
-import io.vertx.ext.web.RoutingContext;
 import io.vertx.ext.web.handler.graphql.VertxDataFetcher;
-import org.apache.kafka.common.config.SaslConfigs;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -20,40 +19,11 @@ public class TopicsDeleteHandler extends CommonHandler {
 
     public static VertxDataFetcher deleteTopics(Map<String, Object> acConfig, Vertx vertx) {
         VertxDataFetcher<List<String>> dataFetcher = new VertxDataFetcher<>((environment, prom) -> {
-            RoutingContext rc = environment.getContext();
-            String token = rc.request().getHeader("Authorization");
-            if (token != null) {
-                if (token.startsWith("Bearer ")) {
-                    token = token.substring("Bearer ".length());
-                }
-                log.info("auth token is {}", token);
-                log.info(SaslConfigs.SASL_JAAS_CONFIG + "is org.apache.kafka.common.security.oauthbearer.OAuthBearerLoginModule required oauth.access.token=\"" + token + "\";");
-                acConfig.put(SaslConfigs.SASL_JAAS_CONFIG, "org.apache.kafka.common.security.oauthbearer.OAuthBearerLoginModule required oauth.access.token=\"" + token + "\";");
-            }
-
-            AdminClientWrapper acw = new AdminClientWrapper(vertx, acConfig);
-            try {
-                acw.open();
-            } catch (Exception e) {
-                log.error(e);
-                if (acw != null) {
-                    acw.close();
-                }
-                prom.fail(e);
-                return;
-            }
+            setOAuthToken(acConfig, environment);
+            AdminClientWrapper acw = createAdminClient(vertx, acConfig, prom);
 
             List<String> topicsToDelete = environment.getArgument("names");
-            acw.deleteTopics(topicsToDelete, res -> {
-                if (res.failed()) {
-                    log.error(res.cause());
-                    prom.fail(res.cause());
-                    acw.close();
-                } else {
-                    prom.complete(topicsToDelete);
-                    acw.close();
-                }
-            });
+            TopicOperations.deleteTopics(acw, topicsToDelete, prom);
         });
         return dataFetcher;
     }
