+++ pom.xml
@@ -15,6 +15,7 @@
         <module>health</module>
         <module>kafka-admin</module>
         <module>graphql</module>
+        <module>common-data-fetchers</module>
     </modules>
 
     <properties>
